Move matching to clauses module
